requirements.txt
mysql.connector
json
base64
re
